tests: move "show_input_warning_banner()" to contest.py to simplify the code a little
